Add AbstractSpi sensor class
